// packages/websocket-client/src/types/events.ts
export enum LoopEvent {
	TYPING = "typing",
	POSTED = "posted",
	POST_EDITED = "post_edited",
	POST_DELETED = "post_deleted",
	POST_UNREAD = "post_unread",
	BURN_ON_READ_POST_REVEALED = "post_revealed",
	BURN_ON_READ_POST_BURNED = "post_burned",
	BURN_ON_READ_POST_ALL_REVEALED = "burn_on_read_all_revealed",
	CHANNEL_CONVERTED = "channel_converted",
	CHANNEL_CREATED = "channel_created",
	CHANNEL_DELETED = "channel_deleted",
	CHANNEL_RESTORED = "channel_restored",
	CHANNEL_UPDATED = "channel_updated",
	CHANNEL_MEMBER_UPDATED = "channel_member_updated",
	CHANNEL_SCHEME_UPDATED = "channel_scheme_updated",
	DIRECT_ADDED = "direct_added",
	GROUP_ADDED = "group_added",
	NEW_USER = "new_user",
	ADDED_TO_TEAM = "added_to_team",
	LEAVE_TEAM = "leave_team",
	UPDATE_TEAM = "update_team",
	DELETE_TEAM = "delete_team",
	RESTORE_TEAM = "restore_team", // This isn't currently used by the web app
	UPDATE_TEAM_SCHEME = "update_team_scheme",
	USER_ADDED = "user_added",
	USER_UPDATED = "user_updated",
	USER_ROLE_UPDATED = "user_role_updated",
	MEMBER_ROLE_UPDATED = "memberrole_updated",
	USER_REMOVED = "user_removed",
	PREFERENCE_CHANGED = "preference_changed",
	PREFERENCES_CHANGED = "preferences_changed",
	PREFERENCES_DELETED = "preferences_deleted",
	EPHEMERAL_MESSAGE = "ephemeral_message",
	STATUS_CHANGE = "status_change",
	HELLO = "hello",
	AUTHENTICATION_CHALLENGE = "authentication_challenge", // This isn't currently used by the web app, and it's a message that would be sent from a client to the server
	REACTION_ADDED = "reaction_added",
	REACTION_REMOVED = "reaction_removed",
	RESPONSE = "response",
	EMOJI_ADDED = "emoji_added",
	MULTIPLE_CHANNELS_VIEWED = "multiple_channels_viewed",
	PLUGIN_STATUSES_CHANGED = "plugin_statuses_changed",
	PLUGIN_ENABLED = "plugin_enabled",
	PLUGIN_DISABLED = "plugin_disabled",
	ROLE_UPDATED = "role_updated",
	LICENSE_CHANGED = "license_changed",
	CONFIG_CHANGED = "config_changed",
	OPEN_DIALOG = "open_dialog",
	GUESTS_DEACTIVATED = "guests_deactivated", // This isn't currently used by the web app
	USER_ACTIVATION_STATUS_CHANGE = "user_activation_status_change",
	RECEIVED_GROUP = "received_group",
	RECEIVED_GROUP_ASSOCIATED_TO_TEAM = "received_group_associated_to_team",
	RECEIVED_GROUP_NOT_ASSOCIATED_TO_TEAM = "received_group_not_associated_to_team",
	RECEIVED_GROUP_ASSOCIATED_TO_CHANNEL = "received_group_associated_to_channel",
	RECEIVED_GROUP_NOT_ASSOCIATED_TO_CHANNEL = "received_group_not_associated_to_channel",
	GROUP_MEMBER_DELETED = "group_member_deleted",
	GROUP_MEMBER_ADDED = "group_member_add",
	SIDEBAR_CATEGORY_CREATED = "sidebar_category_created",
	SIDEBAR_CATEGORY_UPDATED = "sidebar_category_updated",
	SIDEBAR_CATEGORY_DELETED = "sidebar_category_deleted",
	SIDEBAR_CATEGORY_ORDER_UPDATED = "sidebar_category_order_updated",
	CLOUD_SUBSCRIPTION_CHANGED = "cloud_subscription_changed",
	THREAD_UPDATED = "thread_updated",
	THREAD_FOLLOW_CHANGED = "thread_follow_changed",
	THREAD_READ_CHANGED = "thread_read_changed",
	FIRST_ADMIN_VISIT_MARKETPLACE_STATUS_RECEIVED = "first_admin_visit_marketplace_status_received",
	DRAFT_CREATED = "draft_created",
	DRAFT_UPDATED = "draft_updated",
	DRAFT_DELETED = "draft_deleted",
	POST_ACKNOWLEDGEMENT_ADDED = "post_acknowledgement_added",
	POST_ACKNOWLEDGEMENT_REMOVED = "post_acknowledgement_removed",
	PERSISTENT_NOTIFICATION_TRIGGERED = "persistent_notification_triggered",
	HOSTED_CUSTOMER_SIGNUP_PROGRESS_UPDATED = "hosted_customer_signup_progress_updated",
	CHANNEL_BOOKMARK_CREATED = "channel_bookmark_created",
	CHANNEL_BOOKMARK_UPDATED = "channel_bookmark_updated",
	CHANNEL_BOOKMARK_DELETED = "channel_bookmark_deleted",
	CHANNEL_BOOKMARK_SORTED = "channel_bookmark_sorted",
	PRESENCE_INDICATOR = "presence",
	POSTED_NOTIFY_ACK = "posted_notify_ack", // This isn't currently used by the web app
	SCHEDULED_POST_CREATED = "scheduled_post_created",
	SCHEDULED_POST_UPDATED = "scheduled_post_updated",
	SCHEDULED_POST_DELETED = "scheduled_post_deleted",
	CPA_FIELD_CREATED = "custom_profile_attributes_field_created",
	CPA_FIELD_UPDATED = "custom_profile_attributes_field_updated",
	CPA_FIELD_DELETED = "custom_profile_attributes_field_deleted",
	CPA_VALUES_UPDATED = "custom_profile_attributes_values_updated",
	CONTENT_FLAGGING_REPORT_VALUE_UPDATED = "content_flagging_report_value_updated",
	RECAP_UPDATED = "recap_updated",
	POST_TRANSLATION_UPDATED = "post_translation_updated",
	USER_UPDATE_ACTIVE_STATUS = "user_update_active_status",
	GET_STATUSES = "get_statuses",
	GET_STATUSES_BY_IDS = "get_statuses_by_ids"
}